Refactoring of pod template logic
