+++ extensions-contrib/kubernetes-overlord-extensions/src/test/java/org/apache/druid/k8s/overlord/KubernetesTaskRunnerFactoryTest.java
@@ -24,13 +24,10 @@ import com.google.common.base.Supplier;
 import org.apache.druid.indexing.common.TestUtils;
 import org.apache.druid.indexing.common.config.TaskConfig;
 import org.apache.druid.indexing.common.config.TaskConfigBuilder;
-import org.apache.druid.java.util.common.IAE;
 import org.apache.druid.java.util.emitter.service.ServiceEmitter;
 import org.apache.druid.k8s.overlord.common.DruidKubernetesClient;
 import org.apache.druid.k8s.overlord.execution.KubernetesTaskRunnerDynamicConfig;
-import org.apache.druid.k8s.overlord.taskadapter.MultiContainerTaskAdapter;
-import org.apache.druid.k8s.overlord.taskadapter.PodTemplateTaskAdapter;
-import org.apache.druid.k8s.overlord.taskadapter.SingleContainerTaskAdapter;
+import org.apache.druid.k8s.overlord.taskadapter.TaskAdapter;
 import org.apache.druid.server.DruidNode;
 import org.apache.druid.server.log.StartupLoggingConfig;
 import org.apache.druid.tasklogs.NoopTaskLogs;
@@ -40,7 +37,6 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-import java.net.URL;
 import java.util.Properties;
 
 public class KubernetesTaskRunnerFactoryTest
@@ -56,6 +52,7 @@ public class KubernetesTaskRunnerFactoryTest
   private DruidKubernetesClient druidKubernetesClient;
   @Mock private ServiceEmitter emitter;
   @Mock private Supplier<KubernetesTaskRunnerDynamicConfig> dynamicConfigRef;
+  @Mock private TaskAdapter taskAdapter;
 
   @Before
   public void setup()
@@ -87,214 +84,15 @@ public class KubernetesTaskRunnerFactoryTest
         objectMapper,
         null,
         kubernetesTaskRunnerConfig,
-        startupLoggingConfig,
         taskLogs,
-        druidNode,
-        taskConfig,
-        properties,
         druidKubernetesClient,
         emitter,
-        dynamicConfigRef
+        taskAdapter
     );
 
     KubernetesTaskRunner expectedRunner = factory.build();
     KubernetesTaskRunner actualRunner = factory.get();
 
     Assert.assertEquals(expectedRunner, actualRunner);
-  }
-
-  @Test
-  public void test_build_withoutSidecarSupport_returnsKubernetesTaskRunnerWithSingleContainerTaskAdapter()
-  {
-    KubernetesTaskRunnerFactory factory = new KubernetesTaskRunnerFactory(
-        objectMapper,
-        null,
-        kubernetesTaskRunnerConfig,
-        startupLoggingConfig,
-        taskLogs,
-        druidNode,
-        taskConfig,
-        properties,
-        druidKubernetesClient,
-        emitter,
-        dynamicConfigRef
-    );
-
-    KubernetesTaskRunner runner = factory.build();
-
-    Assert.assertNotNull(runner);
-    Assert.assertTrue(runner.adapter instanceof SingleContainerTaskAdapter);
-  }
-
-  @Test
-  public void test_build_withSidecarSupport_returnsKubernetesTaskRunnerWithMultiContainerTaskAdapter()
-  {
-    kubernetesTaskRunnerConfig = KubernetesTaskRunnerConfig.builder()
-        .withCapacity(1)
-        .withSidecarSupport(true)
-        .build();
-
-    KubernetesTaskRunnerFactory factory = new KubernetesTaskRunnerFactory(
-        objectMapper,
-        null,
-        kubernetesTaskRunnerConfig,
-        startupLoggingConfig,
-        taskLogs,
-        druidNode,
-        taskConfig,
-        properties,
-        druidKubernetesClient,
-        emitter,
-        dynamicConfigRef
-    );
-
-    KubernetesTaskRunner runner = factory.build();
-
-    Assert.assertNotNull(runner);
-    Assert.assertTrue(runner.adapter instanceof MultiContainerTaskAdapter);
-  }
-
-  @Test
-  public void test_build_withSingleContainerAdapterType_returnsKubernetesTaskRunnerWithSingleContainerTaskAdapter()
-  {
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.adapter.type", "overlordSingleContainer");
-
-    KubernetesTaskRunnerFactory factory = new KubernetesTaskRunnerFactory(
-        objectMapper,
-        null,
-        kubernetesTaskRunnerConfig,
-        startupLoggingConfig,
-        taskLogs,
-        druidNode,
-        taskConfig,
-        props,
-        druidKubernetesClient,
-        emitter,
-        dynamicConfigRef
-    );
-
-    KubernetesTaskRunner runner = factory.build();
-
-    Assert.assertNotNull(runner);
-    Assert.assertTrue(runner.adapter instanceof SingleContainerTaskAdapter);
-  }
-
-  @Test
-  public void test_build_withSingleContainerAdapterTypeAndSidecarSupport_throwsIAE()
-  {
-    kubernetesTaskRunnerConfig = KubernetesTaskRunnerConfig.builder()
-        .withCapacity(1)
-        .withSidecarSupport(true)
-        .build();
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.adapter.type", "overlordSingleContainer");
-
-    KubernetesTaskRunnerFactory factory = new KubernetesTaskRunnerFactory(
-        objectMapper,
-        null,
-        kubernetesTaskRunnerConfig,
-        startupLoggingConfig,
-        taskLogs,
-        druidNode,
-        taskConfig,
-        props,
-        druidKubernetesClient,
-        emitter,
-        dynamicConfigRef
-    );
-
-    Assert.assertThrows(
-        "Invalid pod adapter [overlordSingleContainer], only pod adapter [overlordMultiContainer] can be specified when sidecarSupport is enabled",
-        IAE.class,
-        factory::build
-    );
-  }
-
-  @Test
-  public void test_build_withMultiContainerAdapterType_returnsKubernetesTaskRunnerWithMultiContainerTaskAdapter()
-  {
-    kubernetesTaskRunnerConfig = KubernetesTaskRunnerConfig.builder()
-        .withCapacity(1)
-        .withSidecarSupport(true)
-        .build();
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.adapter.type", "overlordMultiContainer");
-
-    KubernetesTaskRunnerFactory factory = new KubernetesTaskRunnerFactory(
-        objectMapper,
-        null,
-        kubernetesTaskRunnerConfig,
-        startupLoggingConfig,
-        taskLogs,
-        druidNode,
-        taskConfig,
-        props,
-        druidKubernetesClient,
-        emitter,
-        dynamicConfigRef
-    );
-
-    KubernetesTaskRunner runner = factory.build();
-
-    Assert.assertNotNull(runner);
-    Assert.assertTrue(runner.adapter instanceof MultiContainerTaskAdapter);
-  }
-
-  @Test
-  public void test_build_withMultiContainerAdapterTypeAndSidecarSupport_returnsKubernetesTaskRunnerWithMultiContainerTaskAdapter()
-  {
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.adapter.type", "overlordMultiContainer");
-
-    KubernetesTaskRunnerFactory factory = new KubernetesTaskRunnerFactory(
-        objectMapper,
-        null,
-        kubernetesTaskRunnerConfig,
-        startupLoggingConfig,
-        taskLogs,
-        druidNode,
-        taskConfig,
-        props,
-        druidKubernetesClient,
-        emitter,
-        dynamicConfigRef
-    );
-
-    KubernetesTaskRunner runner = factory.build();
-
-    Assert.assertNotNull(runner);
-    Assert.assertTrue(runner.adapter instanceof MultiContainerTaskAdapter);
-  }
-
-  @Test
-  public void test_build_withPodTemplateAdapterType_returnsKubernetesTaskRunnerWithPodTemplateTaskAdapter()
-  {
-    URL url = this.getClass().getClassLoader().getResource("basePodTemplate.yaml");
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.adapter.type", "customTemplateAdapter");
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", url.getPath());
-
-    KubernetesTaskRunnerFactory factory = new KubernetesTaskRunnerFactory(
-        objectMapper,
-        null,
-        kubernetesTaskRunnerConfig,
-        startupLoggingConfig,
-        taskLogs,
-        druidNode,
-        taskConfig,
-        props,
-        druidKubernetesClient,
-        emitter,
-        dynamicConfigRef
-    );
-
-    KubernetesTaskRunner runner = factory.build();
-
-    Assert.assertNotNull(runner);
-    Assert.assertTrue(runner.adapter instanceof PodTemplateTaskAdapter);
   }
 }
