+++ extensions-contrib/kubernetes-overlord-extensions/src/test/java/org/apache/druid/k8s/overlord/taskadapter/PodTemplateTaskAdapterTest.java
@@ -21,11 +21,8 @@ package org.apache.druid.k8s.overlord.taskadapter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.Optional;
-import com.google.common.base.Supplier;
 import com.google.common.collect.ImmutableMap;
 import io.fabric8.kubernetes.api.model.PodTemplate;
-import io.fabric8.kubernetes.api.model.PodTemplateBuilder;
-import io.fabric8.kubernetes.api.model.VolumeBuilder;
 import io.fabric8.kubernetes.api.model.batch.v1.Job;
 import io.fabric8.kubernetes.api.model.batch.v1.JobBuilder;
 import org.apache.commons.lang.RandomStringUtils;
@@ -35,16 +32,11 @@ import org.apache.druid.indexing.common.config.TaskConfig;
 import org.apache.druid.indexing.common.config.TaskConfigBuilder;
 import org.apache.druid.indexing.common.task.NoopTask;
 import org.apache.druid.indexing.common.task.Task;
-import org.apache.druid.java.util.common.IAE;
 import org.apache.druid.k8s.overlord.KubernetesTaskRunnerConfig;
 import org.apache.druid.k8s.overlord.common.Base64Compression;
 import org.apache.druid.k8s.overlord.common.DruidK8sConstants;
 import org.apache.druid.k8s.overlord.common.K8sTaskId;
 import org.apache.druid.k8s.overlord.common.K8sTestUtils;
-import org.apache.druid.k8s.overlord.execution.DefaultKubernetesTaskRunnerDynamicConfig;
-import org.apache.druid.k8s.overlord.execution.KubernetesTaskRunnerDynamicConfig;
-import org.apache.druid.k8s.overlord.execution.Selector;
-import org.apache.druid.k8s.overlord.execution.SelectorBasedPodTemplateSelectStrategy;
 import org.apache.druid.server.DruidNode;
 import org.apache.druid.server.coordination.BroadcastDatasourceLoadingSpec;
 import org.apache.druid.tasklogs.TaskLogs;
@@ -53,32 +45,25 @@ import org.junit.Assert;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
 import org.mockito.Mockito;
-import org.mockito.internal.util.collections.Sets;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.nio.charset.Charset;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.Collections;
 import java.util.Map;
-import java.util.Properties;
 import java.util.stream.Collectors;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class PodTemplateTaskAdapterTest
 {
-  @TempDir private Path tempDir;
   private KubernetesTaskRunnerConfig taskRunnerConfig;
   private PodTemplate podTemplateSpec;
   private TaskConfig taskConfig;
   private DruidNode node;
   private ObjectMapper mapper;
   private TaskLogs taskLogs;
-  private Supplier<KubernetesTaskRunnerDynamicConfig> dynamicConfigRef;
 
   @BeforeEach
   public void setup()
@@ -98,85 +83,12 @@ public class PodTemplateTaskAdapterTest
     podTemplateSpec = K8sTestUtils.fileToResource("basePodTemplate.yaml", PodTemplate.class);
 
     taskLogs = EasyMock.createMock(TaskLogs.class);
-    dynamicConfigRef = () -> new DefaultKubernetesTaskRunnerDynamicConfig(KubernetesTaskRunnerDynamicConfig.DEFAULT_STRATEGY);
-  }
-
-  @Test
-  public void test_fromTask_withoutBasePodTemplateInRuntimeProperites_raisesIAE()
-  {
-    Exception exception = Assert.assertThrows(
-        "No base prop should throw an IAE",
-        IAE.class,
-        () -> new PodTemplateTaskAdapter(
-        taskRunnerConfig,
-        taskConfig,
-        node,
-        mapper,
-        new Properties(),
-        taskLogs,
-        dynamicConfigRef
-    ));
-    Assert.assertEquals(exception.getMessage(), "Pod template task adapter requires a base pod template to be specified under druid.indexer.runner.k8s.podTemplate.base");
-  }
-
-  @Test
-  public void test_fromTask_withBasePodTemplateInRuntimeProperites_withEmptyFile_raisesIAE() throws IOException
-  {
-    Path templatePath = Files.createFile(tempDir.resolve("empty.yaml"));
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
-
-    Exception exception = Assert.assertThrows(
-        "Empty base pod template should throw a exception",
-        IAE.class,
-        () -> new PodTemplateTaskAdapter(
-        taskRunnerConfig,
-        taskConfig,
-        node,
-        mapper,
-        props,
-        taskLogs,
-        dynamicConfigRef
-    ));
-
-    Assert.assertTrue(exception.getMessage().contains("Failed to load pod template file for"));
-  }
-
-  @Test
-  public void test_fromTask_withBasePodTemplateInRuntimeProperites() throws IOException
-  {
-    Path templatePath = Files.createFile(tempDir.resolve("base.yaml"));
-    mapper.writeValue(templatePath.toFile(), podTemplateSpec);
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
-
-    PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
-        taskRunnerConfig,
-        taskConfig,
-        node,
-        mapper,
-        props,
-        taskLogs,
-        dynamicConfigRef
-    );
-
-    Task task = new NoopTask("id", "id", "datasource", 0, 0, null);
-    Job actual = adapter.fromTask(task);
-    Job expected = K8sTestUtils.fileToResource("expectedNoopJobBase.yaml", Job.class);
-
-    assertJobSpecsEqual(actual, expected);
   }
 
   @Test
   public void test_fromTask_withBasePodTemplateInRuntimeProperites_andTlsEnabled() throws IOException
   {
-    Path templatePath = Files.createFile(tempDir.resolve("base.yaml"));
-    mapper.writeValue(templatePath.toFile(), podTemplateSpec);
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
+    TestPodTemplateSelector podTemplateSelector = new TestPodTemplateSelector(podTemplateSpec);
 
     PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
         taskRunnerConfig,
@@ -191,9 +103,8 @@ public class PodTemplateTaskAdapterTest
             true
         ),
         mapper,
-        props,
         taskLogs,
-        dynamicConfigRef
+        podTemplateSelector
     );
 
     Task task = new NoopTask("id", "id", "datasource", 0, 0, null);
@@ -204,72 +115,17 @@ public class PodTemplateTaskAdapterTest
   }
 
   @Test
-  public void test_fromTask_withNoopPodTemplateInRuntimeProperties_withEmptyFile_raisesIAE() throws IOException
-  {
-    Path baseTemplatePath = Files.createFile(tempDir.resolve("base.yaml"));
-    Path noopTemplatePath = Files.createFile(tempDir.resolve("noop.yaml"));
-    mapper.writeValue(baseTemplatePath.toFile(), podTemplateSpec);
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", baseTemplatePath.toString());
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.noop", noopTemplatePath.toString());
-
-    Assert.assertThrows(IAE.class, () -> new PodTemplateTaskAdapter(
-        taskRunnerConfig,
-        taskConfig,
-        node,
-        mapper,
-        props,
-        taskLogs,
-        dynamicConfigRef
-    ));
-  }
-
-  @Test
-  public void test_fromTask_withNoopPodTemplateInRuntimeProperites() throws IOException
-  {
-    Path templatePath = Files.createFile(tempDir.resolve("noop.yaml"));
-    mapper.writeValue(templatePath.toFile(), podTemplateSpec);
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.noop", templatePath.toString());
-
-    PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
-        taskRunnerConfig,
-        taskConfig,
-        node,
-        mapper,
-        props,
-        taskLogs,
-        dynamicConfigRef
-    );
-
-    Task task = new NoopTask("id", "id", "datasource", 0, 0, null);
-    Job actual = adapter.fromTask(task);
-    Job expected = K8sTestUtils.fileToResource("expectedNoopJob.yaml", Job.class);
-
-    assertJobSpecsEqual(actual, expected);
-  }
-
-  @Test
   public void test_fromTask_withNoopPodTemplateInRuntimeProperites_dontSetTaskJSON() throws IOException
   {
-    Path templatePath = Files.createFile(tempDir.resolve("noop.yaml"));
-    mapper.writeValue(templatePath.toFile(), podTemplateSpec);
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.noop", templatePath.toString());
+    TestPodTemplateSelector podTemplateSelector = new TestPodTemplateSelector(podTemplateSpec);
 
     PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
         taskRunnerConfig,
         taskConfig,
         node,
         mapper,
-        props,
         taskLogs,
-        dynamicConfigRef
+        podTemplateSelector
     );
 
     Task task = new NoopTask(
@@ -290,20 +146,16 @@ public class PodTemplateTaskAdapterTest
   @Test
   public void test_fromTask_withoutAnnotations_throwsDruidException() throws IOException
   {
-    Path templatePath = Files.createFile(tempDir.resolve("base.yaml"));
-    mapper.writeValue(templatePath.toFile(), podTemplateSpec);
+    TestPodTemplateSelector podTemplateSelector = new TestPodTemplateSelector(podTemplateSpec);
 
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
 
     PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
         taskRunnerConfig,
         taskConfig,
         node,
         mapper,
-        props,
         taskLogs,
-        dynamicConfigRef
+        podTemplateSelector
     );
 
     Job job = K8sTestUtils.fileToResource("baseJobWithoutAnnotations.yaml", Job.class);
@@ -313,21 +165,17 @@ public class PodTemplateTaskAdapterTest
   }
 
   @Test
-  public void test_getTaskId() throws IOException
+  public void test_getTaskId()
   {
-    Path templatePath = Files.createFile(tempDir.resolve("base.yaml"));
-    mapper.writeValue(templatePath.toFile(), podTemplateSpec);
+    TestPodTemplateSelector podTemplateSelector = new TestPodTemplateSelector(podTemplateSpec);
 
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
     PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
         taskRunnerConfig,
         taskConfig,
         node,
         mapper,
-        props,
         taskLogs,
-        dynamicConfigRef
+        podTemplateSelector
     );
     Job job = new JobBuilder()
         .editSpec().editTemplate().editMetadata()
@@ -340,19 +188,14 @@ public class PodTemplateTaskAdapterTest
   @Test
   public void test_getTaskId_noAnnotations() throws IOException
   {
-    Path templatePath = Files.createFile(tempDir.resolve("base.yaml"));
-    mapper.writeValue(templatePath.toFile(), podTemplateSpec);
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
+    TestPodTemplateSelector podTemplateSelector = new TestPodTemplateSelector(podTemplateSpec);
     PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
         taskRunnerConfig,
         taskConfig,
         node,
         mapper,
-        props,
         taskLogs,
-        dynamicConfigRef
+        podTemplateSelector
     );
     Job job = new JobBuilder()
         .editSpec().editTemplate().editMetadata()
@@ -365,19 +208,14 @@ public class PodTemplateTaskAdapterTest
   @Test
   public void test_getTaskId_missingTaskIdAnnotation() throws IOException
   {
-    Path templatePath = Files.createFile(tempDir.resolve("base.yaml"));
-    mapper.writeValue(templatePath.toFile(), podTemplateSpec);
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
+    TestPodTemplateSelector podTemplateSelector = new TestPodTemplateSelector(podTemplateSpec);
     PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
         taskRunnerConfig,
         taskConfig,
         node,
         mapper,
-        props,
         taskLogs,
-        dynamicConfigRef
+        podTemplateSelector
     );
     Job job = new JobBuilder()
         .editSpec().editTemplate().editMetadata()
@@ -391,20 +229,15 @@ public class PodTemplateTaskAdapterTest
   @Test
   public void test_toTask_withoutTaskAnnotation_throwsIOE() throws IOException
   {
-    Path templatePath = Files.createFile(tempDir.resolve("base.yaml"));
-    mapper.writeValue(templatePath.toFile(), podTemplateSpec);
-
-    Properties props = new Properties();
-    props.put("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
+    TestPodTemplateSelector podTemplateSelector = new TestPodTemplateSelector(podTemplateSpec);
 
     PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
         taskRunnerConfig,
         taskConfig,
         node,
         mapper,
-        props,
         taskLogs,
-        dynamicConfigRef
+        podTemplateSelector
     );
 
     Job baseJob = K8sTestUtils.fileToResource("baseJobWithoutAnnotations.yaml", Job.class);
@@ -424,20 +257,15 @@ public class PodTemplateTaskAdapterTest
   @Test
   public void test_toTask() throws IOException
   {
-    Path templatePath = Files.createFile(tempDir.resolve("base.yaml"));
-    mapper.writeValue(templatePath.toFile(), podTemplateSpec);
-
-    Properties props = new Properties();
-    props.put("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
+    TestPodTemplateSelector podTemplateSelector = new TestPodTemplateSelector(podTemplateSpec);
 
     PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
         taskRunnerConfig,
         taskConfig,
         node,
         mapper,
-        props,
         taskLogs,
-        dynamicConfigRef
+        podTemplateSelector
     );
 
     Job job = K8sTestUtils.fileToResource("baseJob.yaml", Job.class);
@@ -450,11 +278,7 @@ public class PodTemplateTaskAdapterTest
   @Test
   public void test_toTask_useTaskPayloadManager() throws IOException
   {
-    Path templatePath = Files.createFile(tempDir.resolve("base.yaml"));
-    mapper.writeValue(templatePath.toFile(), podTemplateSpec);
-
-    Properties props = new Properties();
-    props.put("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
+    TestPodTemplateSelector podTemplateSelector = new TestPodTemplateSelector(podTemplateSpec);
 
     Task expected = new NoopTask("id", null, "datasource", 0, 0, ImmutableMap.of());
     TaskLogs mockTestLogs = Mockito.mock(TaskLogs.class);
@@ -467,9 +291,8 @@ public class PodTemplateTaskAdapterTest
         taskConfig,
         node,
         mapper,
-        props,
         mockTestLogs,
-        dynamicConfigRef
+        podTemplateSelector
     );
 
     Job job = K8sTestUtils.fileToResource("expectedNoopJob.yaml", Job.class);
@@ -480,21 +303,15 @@ public class PodTemplateTaskAdapterTest
   @Test
   public void test_fromTask_withRealIds() throws IOException
   {
-    Path templatePath = Files.createFile(tempDir.resolve("noop.yaml"));
-    mapper.writeValue(templatePath.toFile(), podTemplateSpec);
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.noop", templatePath.toString());
+    TestPodTemplateSelector podTemplateSelector = new TestPodTemplateSelector(podTemplateSpec);
 
     PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
         taskRunnerConfig,
         taskConfig,
         node,
         mapper,
-        props,
         taskLogs,
-        dynamicConfigRef
+        podTemplateSelector
     );
 
     Task task = new NoopTask(
@@ -515,21 +332,15 @@ public class PodTemplateTaskAdapterTest
   @Test
   public void test_fromTask_taskSupportsQueries() throws IOException
   {
-    Path templatePath = Files.createFile(tempDir.resolve("noop.yaml"));
-    mapper.writeValue(templatePath.toFile(), podTemplateSpec);
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.queryable", templatePath.toString());
+    TestPodTemplateSelector podTemplateSelector = new TestPodTemplateSelector(podTemplateSpec);
 
     PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
         taskRunnerConfig,
         taskConfig,
         node,
         mapper,
-        props,
         taskLogs,
-        dynamicConfigRef
+        podTemplateSelector
     );
 
     Task task = EasyMock.mock(Task.class);
@@ -554,21 +365,16 @@ public class PodTemplateTaskAdapterTest
   @Test
   public void test_fromTask_withBroadcastDatasourceLoadingModeAll() throws IOException
   {
-    Path templatePath = Files.createFile(tempDir.resolve("noop.yaml"));
-    mapper.writeValue(templatePath.toFile(), podTemplateSpec);
+    TestPodTemplateSelector podTemplateSelector = new TestPodTemplateSelector(podTemplateSpec);
 
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", templatePath.toString());
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.queryable", templatePath.toString());
 
     PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
         taskRunnerConfig,
         taskConfig,
         node,
         mapper,
-        props,
         taskLogs,
-        dynamicConfigRef
+        podTemplateSelector
     );
 
     Task task = EasyMock.mock(Task.class);
@@ -588,98 +394,6 @@ public class PodTemplateTaskAdapterTest
                                           .get(0).getEnv().stream()
                                           .filter(env -> env.getName().equals(DruidK8sConstants.LOAD_BROADCAST_DATASOURCE_MODE_ENV))
                                           .collect(Collectors.toList()).get(0).getValue());
-  }
-
-  @Test
-  public void test_fromTask_withIndexKafkaPodTemplateInRuntimeProperties() throws IOException
-  {
-    Path baseTemplatePath = Files.createFile(tempDir.resolve("base.yaml"));
-    mapper.writeValue(baseTemplatePath.toFile(), podTemplateSpec);
-
-    Path kafkaTemplatePath = Files.createFile(tempDir.resolve("kafka.yaml"));
-    PodTemplate kafkaPodTemplate = new PodTemplateBuilder(podTemplateSpec)
-          .editTemplate()
-          .editSpec()
-          .setNewVolumeLike(0, new VolumeBuilder().withName("volume").build())
-          .endVolume()
-          .endSpec()
-          .endTemplate()
-          .build();
-    mapper.writeValue(kafkaTemplatePath.toFile(), kafkaPodTemplate);
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", baseTemplatePath.toString());
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.index_kafka", kafkaTemplatePath.toString());
-
-    PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
-          taskRunnerConfig,
-          taskConfig,
-          node,
-          mapper,
-          props,
-          taskLogs,
-          dynamicConfigRef
-    );
-
-    Task kafkaTask = new NoopTask("id", "id", "datasource", 0, 0, null) {
-      @Override
-      public String getType()
-      {
-        return "index_kafka";
-      }
-    };
-
-    Task noopTask = new NoopTask("id", "id", "datasource", 0, 0, null);
-    Job actual = adapter.fromTask(kafkaTask);
-    Assert.assertEquals(1, actual.getSpec().getTemplate().getSpec().getVolumes().size(), 1);
-
-    actual = adapter.fromTask(noopTask);
-    Assert.assertEquals(0, actual.getSpec().getTemplate().getSpec().getVolumes().size(), 1);
-  }
-
-  @Test
-  public void test_fromTask_matchPodTemplateBasedOnStrategy() throws IOException
-  {
-    String dataSource = "my_table";
-    Path baseTemplatePath = Files.createFile(tempDir.resolve("base.yaml"));
-    mapper.writeValue(baseTemplatePath.toFile(), podTemplateSpec);
-
-    Path lowThroughputTemplatePath = Files.createFile(tempDir.resolve("low-throughput.yaml"));
-    PodTemplate lowThroughputPodTemplate = new PodTemplateBuilder(podTemplateSpec)
-        .editTemplate()
-        .editSpec()
-        .setNewVolumeLike(0, new VolumeBuilder().withName("volume").build())
-        .endVolume()
-        .endSpec()
-        .endTemplate()
-        .build();
-    mapper.writeValue(lowThroughputTemplatePath.toFile(), lowThroughputPodTemplate);
-
-    Properties props = new Properties();
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.base", baseTemplatePath.toString());
-    props.setProperty("druid.indexer.runner.k8s.podTemplate.lowThroughput", lowThroughputTemplatePath.toString());
-    dynamicConfigRef = () -> new DefaultKubernetesTaskRunnerDynamicConfig(new SelectorBasedPodTemplateSelectStrategy(
-        Collections.singletonList(
-            new Selector("lowThrougput", null, null, Sets.newSet(dataSource)
-            ))));
-
-    PodTemplateTaskAdapter adapter = new PodTemplateTaskAdapter(
-        taskRunnerConfig,
-        taskConfig,
-        node,
-        mapper,
-        props,
-        taskLogs,
-        dynamicConfigRef
-    );
-
-    Task taskWithMatchedDatasource = new NoopTask("id", "id", dataSource, 0, 0, null);
-    Task noopTask = new NoopTask("id", "id", "datasource", 0, 0, null);
-    Job actual = adapter.fromTask(taskWithMatchedDatasource);
-    Assert.assertEquals(1, actual.getSpec().getTemplate().getSpec().getVolumes().size(), 1);
-
-    actual = adapter.fromTask(noopTask);
-    Assert.assertEquals(0, actual.getSpec().getTemplate().getSpec().getVolumes().size(), 1);
   }
 
   private void assertJobSpecsEqual(Job actual, Job expected) throws IOException
