--- extensions-contrib/kubernetes-overlord-extensions/src/test/resources/expectedNoopJobBase.yaml
@@ -1,57 +1,0 @@
-apiVersion: batch/v1
-kind: Job
-metadata:
-  name: "id-3e70afe5cd823dfc7dd308eea616426b"
-  labels:
-    druid.k8s.peons: "true"
-    druid.task.id: "id"
-    druid.task.type: "noop"
-    druid.task.group.id: "id"
-    druid.task.datasource: "datasource"
-
-  annotations:
-    task.id: "id"
-    task.type: "noop"
-    task.group.id: "id"
-    task.datasource: "datasource"
-    task.jobTemplate: base
-spec:
-  activeDeadlineSeconds: 14400
-  backoffLimit: 0
-  ttlSecondsAfterFinished: 172800
-  template:
-    metadata:
-      labels:
-        druid.k8s.peons: "true"
-        druid.task.id: "id"
-        druid.task.type: "noop"
-        druid.task.group.id: "id"
-        druid.task.datasource: "datasource"
-      annotations:
-        task: "H4sIAAAAAAAAAD2MvQ4CIRCE32VqijsTG1qLi7W+wArEbHICrmC8EN7dJf40k/lmJtNQthxgEVPKMGCvXsXgKqnm4x89FTqlKm6MBzw+YCA1nvmm8W4/TQYuxRJeBbZ17cJ3ZhvoSbzShVcu2zLOf9cS7pUl+ANlclrCzr2/AQUK0FqZAAAA"
-        tls.enabled: "false"
-        task.id: "id"
-        task.type: "noop"
-        task.group.id: "id"
-        task.datasource: "datasource"
-        task.jobTemplate: base
-    spec:
-      containers:
-        - command:
-            - sleep
-            - "3600"
-          env:
-            - name: "TASK_DIR"
-              value: "/tmp"
-            - name: "TASK_ID"
-              value: "id"
-            - name: "LOAD_BROADCAST_DATASOURCE_MODE"
-              value: "ALL"
-            - name: "LOAD_BROADCAST_SEGMENTS"
-              value: "false"
-            - name: "TASK_JSON"
-              valueFrom:
-                fieldRef:
-                  fieldPath: "metadata.annotations['task']"
-          image: one
-          name: primary
