+++ extensions-contrib/kubernetes-overlord-extensions/src/test/resources/expectedNoopJobLongIds.yaml
@@ -13,7 +13,7 @@ metadata:
     task.type: "noop"
     task.group.id: "api-issued_kill_wikipedia3_omjobnbc_1000-01-01T00:00:00.000Z_2023-05-14T00:00:00.000Z_2023-05-15T17:03:01.220Z"
     task.datasource: "data_source"
-    task.jobTemplate: noop
+    task.jobTemplate: base
 spec:
   activeDeadlineSeconds: 14400
   backoffLimit: 0
@@ -33,7 +33,7 @@ spec:
         task.type: "noop"
         task.group.id: "api-issued_kill_wikipedia3_omjobnbc_1000-01-01T00:00:00.000Z_2023-05-14T00:00:00.000Z_2023-05-15T17:03:01.220Z"
         task.datasource: "data_source"
-        task.jobTemplate: noop
+        task.jobTemplate: base
     spec:
       containers:
         - command:
