+++ extensions-contrib/kubernetes-overlord-extensions/src/test/resources/expectedNoopJobNoTaskJson.yaml
@@ -13,7 +13,7 @@ metadata:
     task.type: "noop"
     task.group.id: "id"
     task.datasource: "datasource"
-    task.jobTemplate: noop
+    task.jobTemplate: base
 spec:
   activeDeadlineSeconds: 14400
   backoffLimit: 0
@@ -32,7 +32,7 @@ spec:
         task.type: "noop"
         task.group.id: "id"
         task.datasource: "datasource"
-        task.jobTemplate: noop
+        task.jobTemplate: base
     spec:
       containers:
         - command:
