+++ extensions-contrib/kubernetes-overlord-extensions/src/test/resources/expectedNoopPodTemplate.yaml
@@ -1,0 +1,12 @@
+apiVersion: v1
+kind: PodTemplate
+metadata:
+  name: test
+template:
+  spec:
+    containers:
+      - command:
+          - sleep
+          - "3600"
+        image: one
+        name: primary
